requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
mediapipe==0.10.7
Pillow==10.0.1
matplotlib==3.7.2
scikit-learn==1.3.0
tabulate==0.9.0